Moved from Angular to Vue
I chose Angular in the first place for type safety. But Vue does support it as well (in Vue 3.0 the Typescript support will be much better, but that will come out somewhere in 2019). Now there are great plugins for Vue in terms of front end design (Vuetify and many more). These are harder to find in Angular so it will probably be easier and quicker to build in Vue.

Not all type problems are yet resolved (with websocket wrapper)